Move `ASLayoutController` related code out of CoreGraphics+ASConvenience
