+++ AsyncDisplayKit/Details/ASLayoutController.h
@@ -18,14 +18,20 @@ NS_ASSUME_NONNULL_BEGIN
 
 @class ASCellNode;
 
+ASDISPLAYNODE_EXTERN_C_BEGIN
+
 typedef struct {
   CGFloat leadingBufferScreenfuls;
   CGFloat trailingBufferScreenfuls;
 } ASRangeTuningParameters;
 
-FOUNDATION_EXPORT ASRangeTuningParameters const ASRangeTuningParametersZero;
+struct ASDirectionalScreenfulBuffer {
+  CGFloat positiveDirection; // Positive relative to iOS Core Animation layer coordinate space.
+  CGFloat negativeDirection;
+};
+typedef struct ASDirectionalScreenfulBuffer ASDirectionalScreenfulBuffer;
 
-FOUNDATION_EXPORT BOOL ASRangeTuningParametersEqualToRangeTuningParameters(ASRangeTuningParameters lhs, ASRangeTuningParameters rhs);
+ASDISPLAYNODE_EXTERN_C_END
 
 @protocol ASLayoutController <NSObject>
